# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Todo App",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/board/columns": {
      "get": {
        "tags": [
          "board"
        ],
        "summary": "List Columns",
        "operationId": "list_columns_api_v1_board_columns_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardColumnResponse"
                  },
                  "title": "Response List Columns Api V1 Board Columns Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "board"
        ],
        "summary": "Create Column",
        "operationId": "create_column_api_v1_board_columns_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardColumnCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardColumnResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/board/columns/{column_id}": {
      "patch": {
        "tags": [
          "board"
        ],
        "summary": "Update Column",
        "description": "Переименование и перемещение колонки между соседями.",
        "operationId": "update_column_api_v1_board_columns__column_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "column_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Column Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardColumnUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardColumnResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "board"
        ],
        "summary": "Delete Column",
        "description": "Задачи не удаляются вместе с колонкой: column_id обнуляется (SET NULL).",
        "operationId": "delete_column_api_v1_board_columns__column_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "column_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Column Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/files": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Upload File",
        "operationId": "upload_file_api_v1_files_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v1_files_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/files/{attachment_id}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Download File",
        "description": "Доступ даёт либо подписанная ссылка, либо access token владельца.",
        "operationId": "download_file_api_v1_files__attachment_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attachment Id"
            }
          },
          {
            "name": "exp",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Exp"
            }
          },
          {
            "name": "sig",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Sig"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/settings": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Read Settings",
        "operationId": "read_settings_api_v1_notifications_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPrefsSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "notifications"
        ],
        "summary": "Update Settings",
        "operationId": "update_settings_api_v1_notifications_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationPrefsSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPrefsSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/notifications/test": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Send Test",
        "description": "Проверка связи: без неё неверный chat_id обнаружился бы только в дедлайн.",
        "operationId": "send_test_api_v1_notifications_test_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachmentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "url",
          "filename",
          "content_type",
          "size_bytes"
        ],
        "title": "AttachmentResponse"
      },
      "BoardColumnCreate": {
        "properties": {
          "project_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "is_done_column": {
            "type": "boolean",
            "title": "Is Done Column",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "title"
        ],
        "title": "BoardColumnCreate"
      },
      "BoardColumnResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "position": {
            "type": "number",
            "title": "Position"
          },
          "is_done_column": {
            "type": "boolean",
            "title": "Is Done Column"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "title",
          "position",
          "is_done_column"
        ],
        "title": "BoardColumnResponse"
      },
      "BoardColumnUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "is_done_column": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Done Column"
          },
          "before_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Id"
          },
          "after_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "After Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BoardColumnUpdate"
      },
      "Body_upload_file_api_v1_files_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_v1_files_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationPrefsSchema": {
        "properties": {
          "email_enabled": {
            "type": "boolean",
            "title": "Email Enabled",
            "default": true
          },
          "telegram_enabled": {
            "type": "boolean",
            "title": "Telegram Enabled",
            "default": false
          },
          "telegram_chat_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^-?[0-9]{1,32}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Chat Id"
          },
          "lead_time_minutes": {
            "type": "integer",
            "maximum": 10080.0,
            "minimum": 5.0,
            "title": "Lead Time Minutes",
            "default": 60
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "NotificationPrefsSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}